requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pypng==0.20220715.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
